feat: refactor search bar component and replace files
