+++ src/App.css
@@ -1,6 +1,9 @@
+* {
+  margin: 0 auto;
+}
+
 #root {
   max-width: 1280px;
   margin: 0 auto;
-  padding: 2rem;
   text-align: center;
 }
