+++ src/components/AboutPage/About.css
@@ -1,0 +1,3 @@
+.title {
+  color: #FFF;
+}
