+++ src/components/AboutPage/About.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import './About.css';
+
+export default class About extends React.Component {
+  render() {
+    return <h1 className="title">Hello world</h1>;
+  }
+}
